+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/persistence/PersistenceContextDiagnosticsParticipant.java
@@ -158,11 +158,9 @@ public class PersistenceContextDiagnosticsParticipant implements IJavaDiagnostic
         IAnnotation[] typeAnnotations = type.getAnnotations();
 
         // CDI scopes
-        for (IAnnotation annotation : typeAnnotations) {
-            for (String scope : SCOPE_FQ_NAMES) {
-                if (DiagnosticUtils.isMatchedAnnotation(unit, annotation, scope)) {
-                    return true;
-                }
+        for (String scope : SCOPE_FQ_NAMES) {
+            if (DiagnosticUtils.isMatchedAnnotation(unit, typeAnnotations, scope)) {
+                return true;
             }
         }
