+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.test/src/main/java/org/eclipse/lsp4jakarta/jdt/test/persistence/JakartaPersistenceTest.java
@@ -692,427 +692,4 @@ public class JakartaPersistenceTest extends BaseJakartaTest {
         assertJavaCodeAction(codeActionParams, IJDT_UTILS, ca);
     }
 
-    /**
-     * Tests that @PersistenceContext on a field in a plain (unmanaged) Java class
-     * produces a diagnostic.
-     *
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     * §7.6 Container-managed Persistence Contexts: the entity manager lifecycle is managed by the
-     * container and its injection is only valid in container-managed components (EJB, CDI, Servlet).
-     */
-    @Test
-    public void persistenceContextInPlainClass() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInPlainClass.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // Field "em" at 0-based line 11, columns 26-28.
-        // The class bears no CDI scope, EJB, or Servlet annotation — the container
-        // cannot inject an EntityManager here.
-        Diagnostic notInManagedComponentDiag = d(11, 26, 28,
-                                                 "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                 DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, notInManagedComponentDiag);
-    }
-
-    /**
-     * Tests that @PersistenceContext in a CDI managed bean (@ApplicationScoped)
-     * does NOT produce a diagnostic.
-     *
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInCdiBeanValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInCdiBeanValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @ApplicationScoped (CDI managed bean)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext in an EJB (@Stateless session bean)
-     * does NOT produce a diagnostic.
-     *
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInEjbValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInEjbValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @Stateless (EJB session bean)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext in a Jakarta Servlet (@WebServlet + HttpServlet)
-     * does NOT produce a diagnostic.
-     *
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInServletValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInServletValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @WebServlet and extends HttpServlet
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext in a CDI bean with @Dependent scope
-     * does NOT produce a diagnostic.
-     *
-     * All CDI-scoped beans are container-managed. @Dependent is the default CDI scope.
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInDependentBeanValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInDependentBeanValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @Dependent (CDI managed bean)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext in a @WebFilter component does NOT produce a diagnostic.
-     *
-     * @WebFilter is a container-managed servlet component; injection is valid there.
-     *            Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInWebFilterValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInWebFilterValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @WebFilter (servlet component)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext in a @WebListener component does NOT produce a diagnostic.
-     *
-     * @WebListener is a container-managed servlet component; injection is valid there.
-     *              Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInWebListenerValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInWebListenerValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @WebListener (servlet component)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext(type = EXTENDED) in a @Stateful EJB does NOT produce a diagnostic.
-     *
-     * §7.6.3: A container-managed EXTENDED persistence context can only be initiated within a
-     * stateful session bean. @Stateful is the sole valid container type for EXTENDED.
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11810
-     */
-    @Test
-    public void persistenceContextExtendedInStatefulValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextExtendedInStateful.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — EXTENDED context in @Stateful EJB is explicitly valid per §7.6.3
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that @PersistenceContext(type = EXTENDED) in a @Stateless EJB produces a diagnostic.
-     *
-     * §7.6.3: A container-managed EXTENDED persistence context can ONLY be initiated within
-     * a stateful session bean. Using EXTENDED in a @Stateless bean violates the spec.
-     * Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11810
-     */
-    @Test
-    public void persistenceContextExtendedInStatelessInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextExtendedInStateless.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // Field "em" at 0-based line 15, columns 26-28.
-        // PersistenceContextType.EXTENDED is only valid in a @Stateful bean per §7.6.3.
-        Diagnostic extendedInStatelessDiag = d(15, 26, 28,
-                                               "PersistenceContextType.EXTENDED can only be used in a @Stateful EJB session bean.",
-                                               DiagnosticSeverity.Error, "jakarta-persistence", "ExtendedPersistenceContextInNonStatefulBean");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, extendedInStatelessDiag);
-    }
-
-    /**
-     * @PersistenceContext in an @Entity class — invalid; @Entity is NOT a managed injection
-     *                     component. A diagnostic must be produced.
-     *                     Spec §7.6: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextInEntityClassInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInEntityClass.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // field "em" at 0-based line 17, col 26-28
-        Diagnostic notInManagedComponentDiag = d(17, 26, 28,
-                                                 "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                 DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, notInManagedComponentDiag);
-    }
-
-    /**
-     * Plain class with two @PersistenceContext fields — each field must independently
-     * produce a diagnostic.
-     * Spec §7.6: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextMultipleFieldsInPlainClassInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextMultipleFieldsInPlainClass.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // "emA" at 0-based line 11, col 26-29
-        Diagnostic emANotInManagedComponentDiag = d(11, 26, 29,
-                                                    "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                    DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        // "emB" at 0-based line 14, col 26-29
-        Diagnostic emBNotInManagedComponentDiag = d(14, 26, 29,
-                                                    "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                    DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, emANotInManagedComponentDiag, emBNotInManagedComponentDiag);
-    }
-
-    /**
-     * @PersistenceContext(type=EXTENDED) in a @Singleton EJB — invalid per §7.6.3.
-     *                                    EXTENDED context can only be initiated within a @Stateful session bean.
-     */
-    @Test
-    public void persistenceContextExtendedInSingletonInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextExtendedInSingleton.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // "em" at 0-based line 14, col 26-28
-        Diagnostic extendedInSingletonDiag = d(14, 26, 28,
-                                               "PersistenceContextType.EXTENDED can only be used in a @Stateful EJB session bean.",
-                                               DiagnosticSeverity.Error, "jakarta-persistence", "ExtendedPersistenceContextInNonStatefulBean");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, extendedInSingletonDiag);
-    }
-
-    /**
-     * @PersistenceContext(type=EXTENDED) in a CDI @ApplicationScoped bean — invalid per §7.6.3.
-     *                                    EXTENDED is an EJB-only concept; CDI beans cannot use it.
-     */
-    @Test
-    public void persistenceContextExtendedInCdiBeanInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextExtendedInCdiBean.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // "em" at 0-based line 14, col 26-28
-        Diagnostic extendedInCdiBeanDiag = d(14, 26, 28,
-                                             "PersistenceContextType.EXTENDED can only be used in a @Stateful EJB session bean.",
-                                             DiagnosticSeverity.Error, "jakarta-persistence", "ExtendedPersistenceContextInNonStatefulBean");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, extendedInCdiBeanDiag);
-    }
-
-    /** @Stateful EJB with default (TRANSACTION) type — no diagnostic expected. */
-    @Test
-    public void persistenceContextTransactionInStatefulValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextTransactionInStateful.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics: @Stateful is a managed component and default type is TRANSACTION
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /** @Singleton EJB with default (TRANSACTION) type — no diagnostic expected. */
-    @Test
-    public void persistenceContextInSingletonValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextInSingletonValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics: @Singleton is a managed component and TRANSACTION type is not restricted
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that {@code @PersistenceContext} at the <em>type</em> level on a plain (unmanaged)
-     * Java class produces a diagnostic.
-     *
-     * <p>{@code @PersistenceContext} carries {@code @Target({TYPE, METHOD, FIELD})}, so a
-     * type-level annotation is syntactically legal but still violates the spec rule that
-     * injection is only valid in container-managed components.
-     *
-     * <p>Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextOnTypeInPlainClassInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextOnTypeInPlainClass.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // Class name "PersistenceContextOnTypeInPlainClass" is at 0-based line 8,
-        // starting at column 13 (after "public class "), length 36.
-        Diagnostic notInManagedComponentDiag = d(8, 13, 49,
-                                                 "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                 DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, notInManagedComponentDiag);
-    }
-
-    /**
-     * Tests that {@code @PersistenceContext} at the <em>type</em> level on a CDI
-     * {@code @RequestScoped} bean does NOT produce a diagnostic.
-     *
-     * <p>CDI-scoped beans are container-managed; type-level injection is valid there.
-     *
-     * <p>Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextOnTypeInCdiBeanValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextOnTypeInCdiBeanValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @RequestScoped (CDI managed bean)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
-    /**
-     * Tests that {@code @PersistenceContext} at the <em>method</em> level on a plain (unmanaged)
-     * Java class produces a diagnostic.
-     *
-     * <p>Setter injection via {@code @PersistenceContext} is syntactically legal but only valid
-     * inside container-managed components.
-     *
-     * <p>Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextOnMethodInPlainClassInvalid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextOnMethodInPlainClass.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // Method name "setEntityManager" is at 0-based line 13,
-        // starting at column 16 (after "    public void "), length 16.
-        Diagnostic notInManagedComponentDiag = d(13, 16, 32,
-                                                 "@PersistenceContext can only be used in a container-managed component such as a CDI bean, EJB, or Servlet.",
-                                                 DiagnosticSeverity.Error, "jakarta-persistence", "PersistenceContextNotInManagedComponent");
-
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS, notInManagedComponentDiag);
-    }
-
-    /**
-     * Tests that {@code @PersistenceContext} at the <em>method</em> level on an EJB
-     * {@code @Stateless} session bean does NOT produce a diagnostic.
-     *
-     * <p>Setter injection is a standard container-managed injection pattern valid in any
-     * EJB session bean.
-     *
-     * <p>Spec ref: https://jakarta.ee/specifications/persistence/3.0/jakarta-persistence-spec-3.0#a11791
-     */
-    @Test
-    public void persistenceContextOnMethodInStatelessValid() throws Exception {
-        IJavaProject javaProject = loadJavaProject("jakarta-sample", "");
-        IFile javaFile = javaProject.getProject().getFile(
-                                                          new Path("src/main/java/io/openliberty/sample/jakarta/persistence/context/PersistenceContextOnMethodInStatelessValid.java"));
-        String uri = javaFile.getLocation().toFile().toURI().toString();
-
-        JakartaJavaDiagnosticsParams diagnosticsParams = new JakartaJavaDiagnosticsParams();
-        diagnosticsParams.setUris(Arrays.asList(uri));
-
-        // No diagnostics expected — class is annotated with @Stateless (EJB session bean)
-        assertJavaDiagnostics(diagnosticsParams, IJDT_UTILS);
-    }
-
 }
